adding bcrypt to hash a secure password
