+++ app/todo.py
@@ -4,16 +4,7 @@ import os
 
 
 
-def load_users():
-    try:
-        with open('users.json', 'r') as file:
-            return json.load(file)
-    except FileNotFoundError:
-        return {}
-    
-def save_users(users):
-    with open('users.json', 'w') as file:
-        json.dump(users, file, indent=4)
+
 
 def get_todo_file(username):
     return f"{username}_todo_list.json"
